feat: Implement shipping method restrictions and rates management with new controllers, services, and enums

+++ app/Http/Controllers/Api/Shipping/Method/Rate/ShippingMethodRateController.php
@@ -1,18 +1,19 @@
 <?php
 
-namespace App\Http\Controllers\Api\Shipping;
+namespace App\Http\Controllers\Api\Shipping\Method\Rate;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Shipping\Rate\StoreShippingRateRequest;
 use App\Http\Requests\Shipping\Rate\UpdateShippingRateRequest;
 use App\Http\Resources\Shipping\ShippingRateResource;
+use App\Models\ShippingMethod;
 use App\Models\ShippingRate;
 use App\Repositories\ShippingRateRepository;
 use App\Services\Shipping\ShippingRateService;
 use Illuminate\Http\Request;
 use Symfony\Component\HttpFoundation\Response;
 
-class ShippingRateController extends Controller
+class ShippingMethodRateController extends Controller
 {
 
     public function __construct(
@@ -22,7 +23,10 @@ class ShippingRateController extends Controller
     {
     }
 
-    public function index(Request $request) {
+    public function index(ShippingMethod $shippingMethod, Request $request) {
+        $this->shippingRateRepository->setQuery(
+            $shippingMethod->rates()
+        );
         $this->shippingRateRepository->setPagination(true);
         $this->shippingRateRepository->setOrderByColumn(
             $request->get('sort', 'created_at')
@@ -36,26 +40,30 @@ class ShippingRateController extends Controller
         $this->shippingRateRepository->setPage(
             $request->get('page', 1)
         );
-        
+
         return ShippingRateResource::collection(
             $this->shippingRateRepository->findMany()
         );
     }
 
-    public function show(ShippingRate $shippingRate, Request $request) {
+    public function show(ShippingMethod $shippingMethod, ShippingRate $shippingRate, Request $request) {
         $this->shippingRateService->setUser($request->user()->user);
         $this->shippingRateService->setSite($request->user()->site);
-        
+        if (!$shippingRate->shippingMethod->is($shippingMethod)) {
+            return response()->json([
+                'message' => 'Shipping rate does not belong to this shipping method',
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         return new ShippingRateResource(
             $shippingRate
         );
     }
 
-    public function store(StoreShippingRateRequest $request) {
+    public function store(ShippingMethod $shippingMethod, StoreShippingRateRequest $request) {
         $this->shippingRateService->setUser($request->user()->user);
         $this->shippingRateService->setSite($request->user()->site);
-        
-        if (!$this->shippingRateService->createShippingRate($request->validated())) {
+
+        if (!$this->shippingRateService->createShippingRate($shippingMethod, $request->validated())) {
             return response()->json([
                 'message' => 'Error creating shipping rate',
             ], Response::HTTP_UNPROCESSABLE_ENTITY);
@@ -65,10 +73,15 @@ class ShippingRateController extends Controller
         ], Response::HTTP_OK);
     }
 
-    public function update(ShippingRate $shippingRate, UpdateShippingRateRequest $request) {
+    public function update(ShippingMethod $shippingMethod, ShippingRate $shippingRate, UpdateShippingRateRequest $request) {
         $this->shippingRateService->setUser($request->user()->user);
         $this->shippingRateService->setSite($request->user()->site);
 
+        if (!$shippingRate->shippingMethod->is($shippingMethod)) {
+            return response()->json([
+                'message' => 'Shipping rate does not belong to this shipping method',
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         if (!$this->shippingRateService->updateShippingRate($shippingRate, $request->validated())) {
             return response()->json([
                 'message' => 'Error updating shipping rate',
@@ -78,10 +91,15 @@ class ShippingRateController extends Controller
             'message' => 'Shipping rate updated',
         ], Response::HTTP_OK);
     }
-    public function destroy(ShippingRate $shippingRate, Request $request) {
+    public function destroy(ShippingMethod $shippingMethod, ShippingRate $shippingRate, Request $request) {
         $this->shippingRateService->setUser($request->user()->user);
         $this->shippingRateService->setSite($request->user()->site);
-        
+
+        if (!$shippingRate->shippingMethod->is($shippingMethod)) {
+            return response()->json([
+                'message' => 'Shipping rate does not belong to this shipping method',
+            ], Response::HTTP_UNPROCESSABLE_ENTITY);
+        }
         if (!$this->shippingRateService->deleteShippingRate($shippingRate)) {
             return response()->json([
                 'message' => 'Error deleting shipping rate',
